# backend_db/backend/openapi_extracted.yaml
openapi: 3.1.0
info:
  title: SelfDB
  description: Full Self-Hosted BaaS Built for AI Agents.
  version: 0.0.5
paths:
...
  /storage/buckets/count:
    get:
      tags:
      - storage-buckets
      summary: Get Bucket Count
      description: Get total number of buckets accessible to the user.
      operationId: get_bucket_count_storage_buckets_count_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
            pattern: ^[\x20-\x7E]*$
          - type: 'null'
          title: Search
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: integer
                title: Response Get Bucket Count Storage Buckets Count Get
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable (Missing API Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /storage/buckets/:
    post:
      tags:
      - storage-buckets
      summary: Create Bucket
      description: Create a new storage bucket. Requires authentication.
      operationId: create_bucket_storage_buckets__post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BucketCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BucketResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable (Missing API Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
    get:
      tags:
      - storage-buckets
      summary: List Buckets
      description: 'List buckets with optional search and sorting.


        - For unauthenticated users: Only public buckets

        - For authenticated users: Public buckets + buckets owned by the user'
      operationId: list_buckets_storage_buckets__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          maximum: 2147483647
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
            pattern: ^[\x20-\x7E]*$
          - type: 'null'
          title: Search
      - name: sort_by
        in: query
        required: false
        schema:
          enum:
          - created_at
          - updated_at
          - name
          type: string
          default: created_at
          title: Sort By
      - name: sort_order
        in: query
        required: false
        schema:
          enum:
          - asc
          - desc
          type: string
          default: desc
          title: Sort Order
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BucketResponse'
                title: Response List Buckets Storage Buckets  Get
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable (Missing API Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    endpoints__schema__ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    BucketCreate:
      properties:
        name:
          type: string
          maxLength: 63
          minLength: 3
          pattern: ^[a-z0-9][a-z0-9\-]{1,61}[a-z0-9]$
          title: Name
          examples:
          - my-bucket
          - user-uploads
        public:
          type: boolean
          title: Public
          default: false
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
      additionalProperties: false
      type: object
      required:
      - name
      title: BucketCreate
      description: Request model for creating a bucket.
...
    BucketResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        public:
          type: boolean
          title: Public
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        file_count:
          type: integer
          title: File Count
          default: 0
        total_size:
          type: integer
          title: Total Size
          default: 0
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - name
      - public
      - created_at
      - updated_at
      title: BucketResponse
      description: Response model for bucket (matches frontend Bucket interface).
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /users/token